Begin refactoring from Shameless to DRY
Choose the 'else' and '2' cases as being most alike and make the smallest change that
make the first difference the same.

This is the start of DRYing out the strings using a very specific refactoring technique.
In the micro, the plan is
1) Parse some code.
2) Execute the code without using the result. (step is sometimes skipped)
3) Execute the code and use the result
4) Delete unused code.

In the macro, the plan is
1) find the strings that are most alike
2) find the smallest difference
3) make the smallest change that will remove the difference
4) make changes only on a single line of code
5) run the tests after each change
6) if the tests fail, undo and make a better change

A few things to especially note:
We never chase after red; if the tests fail, undo and make a change that leaves the tests green.
This can be a real challenge but it's a great learning experience and is a technique well
worth knowing.

The rule is to make small differences the same, NOT to make the big commonalities
DRY. Don't get distracted by strings that are the _same_, instead identify and remove
strings that are different.

Do complete 'horizontal' refactorings before switching to a 'vertical' path. This means that
once you pick two strings and start making them the same, you should finish this entirely
before getting distracted by a third string, no matter how much that other string calls out to you.