First step in adding argument to #container
We're gonna need to pass an argument.  Since I can't simultaneously change the caller and
the method signature (this violates the 'changes only on a single line' rule) I have to first
add the argument and set a default.  This dummy default clearly states that it will send me
to the false branch of any if statements that depend on it.

Ok, I know that there _aren't_ any if statements yet, but the next commit will add one.