Add #amount, take 1; start by returning 'no more'
We'll add the #amount method twice, using two different techniques.

When students try to make 'no more' and '#{number-1}' the same, they often unconsciously
do something slightly different then previously.  They seem confused by the fact that 'no more'
is a straightforward string and '#{number-1}' is interpolated, and they often choose to
return 'no more' in #amount.  They don't recognize that they've picked the non-default value,
and in a minute things break for them in a confusing way.  Watch, it's about to happen. :-)