Change private method arg name from 'number' to 'bottle_number'.
What does this argument represent?  A verse?  Or a bottle number?
While it's true that in the #verse method the 'number' argument represents a verse
number, by the time the arg is passed to the private methods it represents a bottle
number.

We know this because the #successor method is used to change the value
of the number that gets passed to #amount and #container.  During any verse N,
#amount might get called with (N), or with (N+1).  Therefore, #amount is not taking
verse number as an argument, it's taking bottle number.

Naming the concept correctly makes everything clear.  We should next extract a
class that holds bottle_number as a piece of state, and move methods that depend
on that piece of state to it.