Add initializer and attr_reader on 'number' to BottleNumber.
This is a bit of new code that we need in BottleNumber.  The extract class refactoring
tells us to save this bit of state in a BottleNumber instance and change all of the methods
to refer to this bit of state instead of taking an argument.  This is the first step in that direction.

Still not using this code, just making sure it compiles.