Change all 'bottle_number' to 'number' in BottleNumber class.
The bit of state is named 'number' but when all of the methods in BottleNumber still
refer to the name we used when they were part of the Bottle class.  Now that BottleNumber
is holding on to state in 'number', and now that I've had to default one of the method args
to 'self.number', I can see that I'd like all the methods of BottleNumber to refer to 'number'
instead of 'bottle_number'.  I'll allow myself a global change to fix these names.