Create empty BottleNumber1 subclass.
See 'Replace Conditional with Polymorphism' in Refactoring to Patterns (Joshua Kerievsky)
and Refactoring (Martin Fowler). I'm gonna refactor to subclasses here although we may
slightly change our minds about of this later.

Notice that the BottleNumber class contains methods all of the same shape.
Each method contains an 'if' statement that switches on a piece of state where each
branch of the if statement returns the most atomic value for the combination of the current
bit of state and the concept named by the method.

The methods contain NO code other than the 'if' statements and each branch of each
if statement returns the smallest, driest thing related to the concept it represents.

Once we get the code into this shape (which only happened because we followed a strict
set of refactoring rules) we can begin to create a separate class for each type VALUE
that we switch on.

There are things we don't like about the current code but we'll put off fixing them until later.
1) Bottles is a terrible class name
2) bottle_number.successor should return an instance of BottleNumber
    instead of an instance of Fixnum.

I'm bothered by both of those things, but I want to stay the course here, i.e., continue
the horizontal refactoring through to the end, before I switch to a vertical idea.  This means
not yet changing how #successor behaves, even though I don't like it.