Dry out getting a new BottleNumber.
I can see that I need to sometimes get an instance of BottleNumber and sometimes
get an instance of BottleNumber1.  This will likely involve a condition and I don't want
to put the conditional in two places, so I'll create a new method to get instances of
the correct variant of BottleNumber.  Right now it only needs to mimic the current
behavior.

This tests parsing only.