Change #bottle_number_for to return an instance of the correct class.
This change make #bottle_number_for a factory method that uses the value of 'number'
to choose the correct class.

Notice that the structure of this if statement mirrors the structure of the if statement
in BottleNumber#container.  Now, instead of switching on the value of type to know what
to do we are switching on the value of type to know what class to use.

We used to both know why we switched and what to do when we did, and this change
reduces our knowledge to why we switch and the name of the class that knows what
to do.

This is a huge change and this transition is at the core of using many small objects.
Small objects are easy to reason about, reuse and test.  This code is simpler, but there's
a lot of simple code.  The problem has changed for knowing what to do to knowing what
object to get.