Copy the factory code to BottleNumber#for.
1) The #bottle_number_for method clearly doesn't belong in Bottles.  Where should it go?
2) Also, I'm still bothered by #successor returning a Fixnum, when I expect it to return the
next one of the kind of thing it is, like all other successor methods.

Are these two issues related?  Let's explore.

Do we have an existing class into which it makes sense to put the factory?  Let's put
it into BottleNumber and see if we like it.

This just parses the code.