Way too big a change to illustrate changing the contract with #successor.
I want to fix #successor to return an instance of BottleNumberx.  This means it has to
use the factory.  This code I've just checked in works, but it makes changes in three
unrelated places.  I've checked it in just to illustrate where we're going, but I'll next
back it out and take us there step by step.