Add temporary trap for BottleNumber#for being passed a BottleNumber.
I'll change methods one at a time to make #success return a BottleNumber instead of a
Fixnum.  Until I get completely through the transition I might get either kind of thing as
an argument to the factory method.

If the factory gets a BottleNumber, it'll just return it, otherwise it'll create one.