Refine Fixnum with to_bottle_number.
Right now, to get the correct kind of BottleNumberX you need to know three things.
1) The Fixnum for which you want a bottle number.
2) The class of the factory (BottleNumber).
3) The message to send to the factory to get the correct instance (#for).

How can we reduce this to just two things?
Well, you cannot avoid knowing the Fixnum and you have to send someone a message,
so the only bit of knowledge that we can get rid of is having to know the name of the
class of the factory.

We don't say
StringUtils.concat( 'stringa', 'stringb')
we say
'stringa'.concat('stringb')

And to convert an integer to a string, we don't say
IntegerConverter.convert_to_string(14)
we say
14.to_s

I want to send a message to the object I already have.  What message should I send
a Fixnum to ask it to turn itself into a BottleNumber? #to_bottle_number.

It's hard to think of this when we imagine primitive types some special, unchangeable
thing, but they are just objects.