Refactor the DRY house to ease RandomHouse.
This refactoring changes a few methods to have one responsibility, and it
replaces '' with 'the house that Jack built'.

One could argue that 'the house that Jack built' _is_ correct, and this makes
RandomHouse easier.  It feels more important to teach them to proceed
from here, than to teach them how to get here from house_2_shameles_to_dry.