Random via primitive obsession on a boolean.
This breaks all of the refactoring rules.  It's just a spike to explore the problem,
and it has the advantage of expressing the problem in a form we've seen before.
The 'if' statement in #initialize looks a lot like those methods we wrote for 99
Bottles.

If only that boolean were smarter......