Add 'if' statement that switches on 'animal.nil?'
We've spent the entire course talking about removing conditionals, and here we've just
added one.
In the prior lessons on removing conditionals, we've learned that the thing we're switching
on should be turned into an object, and the object should be given the behavior.

We need to create an object to stand in for 'animal.nil?' and this object needs to implement
#species and #sound to return '<silence>'.