Various variations on House
Sorry, step-by-step commits, this is the end point of a refactoring.

Notice that the data has been extracted into the HousePhrases class which
now takes responsibility for providing House with a list of the right things
in the right order.

FixedLastRandomOrder and FixedLastRandomOrder are new, and can be injected
into HousePhrases to vary the order.

I haven't updated the tests, but imagine how testing this would be.....